add url based navigation (in front with #)
