+++ frontend/package.json
@@ -23,6 +23,7 @@
     "chart.js": "^4.5.1",
     "chartjs-adapter-date-fns": "^3.0.0",
     "chartjs-plugin-zoom": "^2.2.0",
-    "hammerjs": "^2.0.8"
+    "hammerjs": "^2.0.8",
+    "svelte-spa-router": "^4.0.1"
   }
 }
