+++ frontend/src/routes/LogsView.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+  import LogsPage from '../lib/LogsPage.svelte';
+</script>
+
+<LogsPage />
