+++ static/index.html
@@ -5,8 +5,8 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Home Assistant</title>
-    <script type="module" crossorigin src="/assets/index-BoTXoESY.js"></script>
-    <link rel="stylesheet" crossorigin href="/assets/index-BFKrJ6Ay.css">
+    <script type="module" crossorigin src="/assets/index-CPGf1T6f.js"></script>
+    <link rel="stylesheet" crossorigin href="/assets/index-cxc8MmNM.css">
   </head>
   <body>
     <div id="app"></div>
